Improve build and file size
This upgrades to `rollup@0.41.4` which fixes issues with ES3 Babel
transforms and Babel helpers which has reduced the build file size.

File | Before | After | Delta
---- | ------ | ----- | -----
react-intl.min.js | 47.10 kb | 41.61 kb | -11.66%
react-intl.min.js.gz | 13.31 kb | 11.51 kb | -13.52%

Fixes #814
Relates #788, #687
