+++ .babelrc
@@ -1,10 +1,12 @@
 {
-    "plugins": [
-        "transform-es2015-modules-commonjs"
-    ],
-    "ignore": [
-        "dist/*.js",
-        "lib/index.js",
-        "locale-data/*.js"
-    ]
+    "env": {
+        "test": {
+            "plugins": [
+                "transform-es2015-modules-commonjs"
+            ],
+            "only": [
+                "lib/index.es.js"
+            ]
+        }
+    }
 }
