+++ package.json
@@ -76,16 +76,15 @@
     "babel-cli": "^6.2.0",
     "babel-eslint": "^7.1.1",
     "babel-jest": "^18.0.0",
+    "babel-plugin-external-helpers": "^6.18.0",
     "babel-plugin-react-intl": "^2.0.0",
     "babel-plugin-transform-class-properties": "^6.11.5",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.4.0",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.18.0",
     "babel-plugin-transform-es3-member-expression-literals": "^6.3.13",
     "babel-plugin-transform-es3-property-literals": "^6.3.13",
     "babel-plugin-transform-object-rest-spread": "^6.1.18",
     "babel-plugin-transform-react-remove-prop-types": "^0.2.10",
-    "babel-polyfill": "^6.3.14",
     "babel-preset-es2015": "^6.1.18",
-    "babel-preset-es2015-rollup": "^3.0.0",
     "babel-preset-react": "^6.1.18",
     "babelify": "^7.2.0",
     "benchmark": "^2.1.0",
@@ -107,7 +106,7 @@
     "react-addons-test-utils": "^15.0.0",
     "react-dom": "^15.0.0",
     "rimraf": "^2.4.2",
-    "rollup": "^0.41.1",
+    "rollup": "^0.41.4",
     "rollup-plugin-babel": "^2.3.9",
     "rollup-plugin-commonjs": "^7.0.0",
     "rollup-plugin-memory": "^2.0.0",
@@ -121,9 +120,9 @@
   "scripts": {
     "clean": "rimraf src/en.js coverage/ dist/ lib/ locale-data/",
     "build:data": "babel-node scripts/build-data",
-    "build:lib": "babel-node scripts/build-lib",
-    "build:dist:dev": "cross-env NODE_ENV=development babel-node scripts/build-dist",
-    "build:dist:prod": "cross-env NODE_ENV=production babel-node scripts/build-dist",
+    "build:lib": "rollup -c rollup.config.lib.js",
+    "build:dist:dev": "cross-env NODE_ENV=development rollup -c rollup.config.dist.js",
+    "build:dist:prod": "cross-env NODE_ENV=production rollup -c rollup.config.dist.js",
     "build:dist": "npm run build:dist:dev && npm run build:dist:prod",
     "build": "npm run build:data && npm run build:lib && npm run build:dist",
     "react:clean": "rimraf node_modules/{react,react-dom,react-addons-test-utils}",
