+++ src/.babelrc
@@ -1,6 +1,8 @@
 {
     "presets": [
-        "es2015",
+        ["es2015", {
+            "modules": false
+        }],
         "react"
     ],
     "plugins": [
@@ -8,6 +10,25 @@
         "transform-class-properties",
         "transform-es3-member-expression-literals",
         "transform-es3-property-literals",
-        ["transform-react-remove-prop-types", {"mode": "wrap"}]
-    ]
+        ["transform-react-remove-prop-types", {
+            "mode": "wrap"
+        }]
+    ],
+    "env": {
+        "development": {
+            "plugins": [
+                "external-helpers"
+            ]
+        },
+        "production": {
+            "plugins": [
+                "external-helpers"
+            ]
+        },
+        "test": {
+            "plugins": [
+                "transform-es2015-modules-commonjs"
+            ]
+        }
+    }
 }
